"Implement basic Spring Security configuration and views"
This commit includes a setup for basic Spring Security configuration and multiple views for different user roles.

We have added `SpringSecurityConfiguration` and `SecurityConfigInitializer` classes to set up an in-memory authentication with a single user and to create a Security Filter Chain, respectively.

In addition, we built three Controller methods in the `LoginController` class that return different views based on the roles of users: one for all employees, one for HR staff, and one for managers. The corresponding JSP view files `view_for_all_employees.jsp`, `view_for_hr.jsp`, and `view_for_managers.jsp` are added under src/main/webapp/WEB-INF/view.

The Spring MVC components are configured in the `SpringWebConfiguration` class. A `ViewResolver` bean is created to map view names to actual views.

Finally, we updated .gitignore files to ignore Intellij IDEA, Eclipse, Netbeans, VS Code files, build directories and Mac OS .DS_Store files. This was necessary to keep the repository clean from OS and IDE specific files which are not needed in the repository.

Made these changes to facilitate role-based page navigation and to provide a foundation upon which more advanced security configurations can be built.