"helper function was added to streamline the process of adding users to the in-memory authentication store. The password encoder was upgraded to use BCrypt's $2A version, introducing better security.
Part of the HTTP requests authorization was configured based on user roles and specific routes were created for different roles, granting access to specific endpoints within the application.

Next, the `LoginController` was renamed and moved to a new package to better organize the code and `@GetMapping` annotations were replaced with `@RequestMapping` for flexibility in HTTP methods.

Additionally, logging has been enabled in the SpringWebConfiguration class for improved monitoring by adding a new PropertySource that points to 'logging.properties'. Necessary dependencies have been added to the pom.xml for this purpose.

Lastly, an explanatory README.md file was added for developers to understand this project better. Spare web pages and xml file that were no longer necessary or used were deleted. With these changes, code organization improves and overall security strengthens."