"Implement JDBC authentication for improved security
Shifted from in-memory user details to JDBC authentication using a PostgreSQL database. The transition to JDBC enhances security and reliability while facilitating easier management of user details. Introduced necessary dependencies for database connection in pom.xml and outlined DataSource configurations in SpringWebConfiguration.java and dataSources.xml. Additionally, created a schema.sql file as a template to create a database that will be compatible with app."